PIM-3595: move cursor classes in storage utils component
