+++ src/Akeneo/Component/StorageUtils/Cursor/CursorFactoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Bundle\StorageUtilsBundle\Cursor;
+namespace Akeneo\Component\StorageUtils\Cursor;
 
 /**
  * Interface CursorFactoryInterface
