+++ src/Akeneo/Component/StorageUtils/Cursor/CursorInterface.php
@@ -1,5 +1,5 @@
 <?php
-namespace Akeneo\Bundle\StorageUtilsBundle\Cursor;
+namespace Akeneo\Component\StorageUtils\Cursor;
 
 /**
  * Interface CursorInterface
