+++ src/Akeneo/Component/StorageUtils/Cursor/PaginatorFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Bundle\StorageUtilsBundle\Cursor;
+namespace Akeneo\Component\StorageUtils\Cursor;
 
 /**
  * Class PaginatorFactory to instantiate paginator to iterate page of entities
