+++ src/Pim/Bundle/CatalogBundle/Query/ProductQueryBuilderFactory.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Query;
 
-use Akeneo\Bundle\StorageUtilsBundle\Cursor\CursorFactoryInterface;
+use Akeneo\Component\StorageUtils\Cursor\CursorFactoryInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use Pim\Bundle\CatalogBundle\Query\Filter\FilterRegistryInterface;
 use Pim\Bundle\CatalogBundle\Query\Sorter\SorterRegistryInterface;
